+++ readium/shared/r2-shared/src/main/java/org/readium/r2/shared/publication/ContentProtection.kt
@@ -9,6 +9,9 @@
 
 package org.readium.r2.shared.publication
 
+import androidx.annotation.StringRes
+import org.readium.r2.shared.R
+import org.readium.r2.shared.UserException
 import org.readium.r2.shared.fetcher.Fetcher
 import org.readium.r2.shared.publication.asset.PublicationAsset
 import org.readium.r2.shared.publication.services.ContentProtectionService
@@ -76,4 +79,35 @@ interface ContentProtection {
         val onCreatePublication: Publication.Builder.() -> Unit = {}
     )
 
+    /**
+     * Represents a specific Content Protection technology, uniquely identified with an [uri].
+     */
+    class Scheme(
+        val uri: String,
+        val name: LocalizedString?
+    ) {
+        override fun hashCode(): Int = uri.hashCode()
+        override fun equals(other: Any?): Boolean = (other as? Scheme)?.uri == uri
+
+        companion object {
+            /** Readium LCP DRM scheme. */
+            val Lcp = Scheme(uri = "http://readium.org/2014/01/lcp", name = LocalizedString("Readium LCP"))
+            /** Adobe ADEPT DRM scheme. */
+            val Adept = Scheme(uri = "http://ns.adobe.com/adept", name = LocalizedString("Adobe ADEPT"))
+        }
+    }
+
+    sealed class Exception(userMessageId: Int, vararg args: Any?, quantity: Int? = null, cause: Throwable? = null) : UserException(userMessageId, quantity, *args, cause = cause) {
+        constructor(@StringRes userMessageId: Int, vararg args: Any?, cause: Throwable? = null) : this(userMessageId, *args, quantity = null, cause = cause)
+
+        /**
+         * Exception returned when the given Content Protection [scheme] is not supported by the
+         * app.
+         */
+        class SchemeNotSupported(val scheme: Scheme? = null) : Exception(
+            if (scheme?.name == null) R.string.r2_shared_publication_content_protection_exception_not_supported_unknown
+            else R.string.r2_shared_publication_content_protection_exception_not_supported,
+            scheme?.name?.string
+        )
+    }
 }